packaging BA-agent
