+++ pyproject.toml
@@ -1,0 +1,41 @@
+[build-system]
+requires = ["hatchling >= 1.26"]
+build-backend = "hatchling.build"
+
+[project]
+name = "batteryarchive_agent"
+version = "0.0.1"
+authors = [ {name = "Lara Kassabian", email=""}, 
+            {name = "Irving Derin", email=""}, 
+            {name = "Valerio De Angelis", email=""},
+            {name = "Yuliya Preger", email=""}]
+description = "A repository for easy visualization, analysis, and comparison of battery data across institutions."
+readme = {file = "README.md", content-type = "text/markdown"}
+requires-python = ">=3.11"
+classifiers = [
+    "Programming Language :: Python :: 3",
+    "Operating System :: OS Independent"
+]
+license = { file = "LICENSE.txt" }
+dependencies = [
+    "matplotlib",
+    "openpyxl",
+    "pandas",
+    "psycopg2-binary",
+    "xlrd",
+    "pyyaml", 
+    "configparser",
+    "pyarrow",
+    "requests",
+    "numpy",
+    "sqlalchemy",
+    "h5py",
+    "click"
+]
+
+[project.urls]
+Homepage = "https://github.com/battery-lcf/batteryarchive-agent"
+Issues = "https://github.com/battery-lcf/batteryarchive-agent/issues"
+
+[tool.hatch.build.targets.wheel]
+packages = ["src/batteryarchive-agent"]
